Create README.md
This game is thrilling and enjoyable that you can take on if you are ready to take a chance. Playing a few times can increase your chances of winning and if you are new, participate No worries! Visit our website to find many helpful tips and tricks that help you determine your probabilities.